Move handle_message and MySQL stuff to lib/commands.js
This makes it easier to cover everything, and it's also probably good practice to not have long functions in server.js
