+++ lib/commands.js
@@ -1,3 +1,5 @@
+const Secrets = require("../secrets");
+
 const basic_math = require("mathjs");
 const math = basic_math.create(basic_math.all);
 math.import({
@@ -7,10 +9,23 @@ math.import({
 	derivative: function () { throw new ReferenceError("Function derivative is disabled"); }
 }, { override: true });
 
+const { promisify } = require("util");
+
+const mysql = require("mysql");
+const sql_pool = mysql.createPool({
+	connectionLimit: 5,
+	host: "localhost",
+	user: "avabur-bot",
+	password: Secrets.sql_pass,
+	database: "avabur"
+});
+sql_pool.query = promisify(sql_pool.query);
+
 const get_currency_prices = require("./get-currency-prices");
+const get_version = require("./get-version.js");
 const {add_commas, remove_commas, expand_numeric_literals} = require("./util");
 
-exports.market = async function (str) {
+async function market (str) {
 	"use strict";
 	let tags = str.split(" ");
 	let currencies = [];
@@ -70,9 +85,9 @@ exports.market = async function (str) {
 		}
 	}
 	return reply.replace(/, $/, "");
-};
+}
 
-exports.calculate = async function (str) {
+async function calculate (str) {
 	if (typeof str === "undefined") {
 		str = "";
 	}
@@ -137,4 +152,65 @@ exports.calculate = async function (str) {
 		}
 	}
 	return reply;
+}
+
+async function handle_message (msg) {
+	if (msg.content === "Everyone An event is starting soon!" && msg.author.bot) {
+		console.log("Received Event starting message");
+		await sql_pool.query("insert into events (time) values (current_timestamp())");
+		console.log("Logged current time in events table");
+	}
+	let reply = "";
+	if (/^![a-zA-Z]+/.test(msg.content)) {
+		switch (msg.content.match(/^![a-zA-Z]+/)[0]) {
+		case "!ping":
+			reply = "pong";
+			break;
+		case "!luck":
+			console.log("Calculating current luck.");
+			try {
+				let result = await sql_pool.query("select unix_timestamp(time) from events;");
+				result = result.map(x => x["unix_timestamp(time)"]);
+				let changes = result.map((x, i, a) => (i > 0 ? x - a[i - 1] : 0));
+				changes.shift();
+
+				// Remove long times where NoA script was down
+				let counted_entries = 0;
+				let avg = changes.reduce((total, x) => { counted_entries++; return total + (x > 21600 ? 0 : x); }, 0) / counted_entries;
+				let time_since_last = Date.now() / 1000 - result[result.length - 1];
+				reply = `Event luck is at ${(time_since_last / avg * 100).toFixed(2)}%.`;
+			} catch (e) {
+				console.error(e.message);
+				reply = "Error calculating luck";
+			}
+			break;
+		case "!market":
+			console.log("Getting market currency values.");
+			reply = await market(msg.content.replace(/^!market ?/, ""));
+			break;
+		case "!source":
+			reply = "avabur-bot by extrafox45#9230 https://github.com/bobpaw/avabur-bot";
+			break;
+		case "!math": case "!calc": case "!calculate":
+			reply = await calculate(msg.content.replace(/^!math ?|!calc(?:ulate)? ?/, ""));
+			break;
+		case "!version":
+			try {
+				reply = await get_version();
+			} catch (e) {
+				console.log("Error getting version: %s", e.message);
+				reply = "Error getting version.";
+			}
+			break;
+		case "!help": case "!commands": default:
+			reply = "!luck, !market, !ping, !source, !version, !help, !commands, !math, !calc, !calculate";
+		}
+	}
+	return reply;
+}
+
+module.exports = {
+	market,
+	calculate,
+	handle_message
 };
